requirements.txt
pymongo
python-dotenv
qrcode
pillow
